Added time displays
Added a new user setting, "Display Local and Track Time", and two new labels to the toolbar at the top of the form, just to the left of the Full Screen button. These will display the track's local time and the user's local time, so any difference in time zone can be noted.