Nullable Object error on new install or patch
The default value for the UseLowScreenResolutionSizes was set to true, but no default font was defined. This caused an error when the user tried to load the leaderboard if they didn't edit and save their user settings first.

Updated the default value for UseLowScreenResolutionSizes to be false, and added default values for the override font in the default user settings file that gets created on first run.